+++ notify-impl/src/main/java/com/dwarfeng/notify/impl/dao/preset/RouterInfoPresetCriteriaMaker.java
@@ -2,14 +2,12 @@ package com.dwarfeng.notify.impl.dao.preset;
 
 import com.dwarfeng.notify.stack.service.RouterInfoMaintainService;
 import com.dwarfeng.subgrade.sdk.hibernate.criteria.PresetCriteriaMaker;
-import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
 import org.hibernate.criterion.DetachedCriteria;
 import org.hibernate.criterion.MatchMode;
 import org.hibernate.criterion.Restrictions;
 import org.springframework.stereotype.Component;
 
 import java.util.Arrays;
-import java.util.Objects;
 
 @Component
 public class RouterInfoPresetCriteriaMaker implements PresetCriteriaMaker {
@@ -17,12 +15,6 @@ public class RouterInfoPresetCriteriaMaker implements PresetCriteriaMaker {
     @Override
     public void makeCriteria(DetachedCriteria detachedCriteria, String s, Object[] objects) {
         switch (s) {
-            case RouterInfoMaintainService.CHILD_FOR_NOTIFY_SETTING:
-                childForNotifySetting(detachedCriteria, objects);
-                break;
-            case RouterInfoMaintainService.CHILD_FOR_NOTIFY_SETTING_ENABLED:
-                childForNotifySettingEnabled(detachedCriteria, objects);
-                break;
             case RouterInfoMaintainService.TYPE_EQUALS:
                 typeEquals(detachedCriteria, objects);
                 break;
@@ -31,34 +23,6 @@ public class RouterInfoPresetCriteriaMaker implements PresetCriteriaMaker {
                 break;
             default:
                 throw new IllegalArgumentException("无法识别的预设: " + s);
-        }
-    }
-
-    @SuppressWarnings("DuplicatedCode")
-    private void childForNotifySetting(DetachedCriteria detachedCriteria, Object[] objects) {
-        try {
-            if (Objects.isNull(objects[0])) {
-                detachedCriteria.add(Restrictions.isNull("notifySettingLongId"));
-            } else {
-                LongIdKey longIdKey = (LongIdKey) objects[0];
-                detachedCriteria.add(Restrictions.eqOrIsNull("notifySettingLongId", longIdKey.getLongId()));
-            }
-        } catch (Exception e) {
-            throw new IllegalArgumentException("非法的参数:" + Arrays.toString(objects));
-        }
-    }
-
-    private void childForNotifySettingEnabled(DetachedCriteria detachedCriteria, Object[] objects) {
-        try {
-            if (Objects.isNull(objects[0])) {
-                detachedCriteria.add(Restrictions.isNull("notifySettingLongId"));
-            } else {
-                LongIdKey longIdKey = (LongIdKey) objects[0];
-                detachedCriteria.add(Restrictions.eqOrIsNull("notifySettingLongId", longIdKey.getLongId()));
-            }
-            detachedCriteria.add(Restrictions.eq("enabled", true));
-        } catch (Exception e) {
-            throw new IllegalArgumentException("非法的参数:" + Arrays.toString(objects));
         }
     }