+++ CHANGELOG.md
@@ -31,7 +31,9 @@
 
 #### 功能移除
 
-- (无)
+- 删除实体字段。
+  - com.dwarfeng.notify.stack.bean.entity.RouterInfo.notifySettingKey。
+  - com.dwarfeng.notify.stack.bean.entity.RouterInfo.enabled。
 
 ---